Update collection commands to use the unified styling, validation, presenters and error handling
